Hotfix - renaming and reordering namespaces.
