+++ .gitignore
@@ -5,3 +5,4 @@ riderModule.iml
 /_ReSharper.Caches/
 .idea/
 MetaExchangeSowaLabs.sln.DotSettings.user
+MetaExchangeSowaLabs.sln.DotSettings.user/
