+++ MetaExchangeSowaLabs.Services/Services/MetaExchangeService.cs
@@ -5,7 +5,8 @@ using MetaExchangeSowaLabs.Core.CustomErrors;
 using MetaExchangeSowaLabs.Core.Entities;
 using MetaExchangeSowaLabs.Core.Enums;
 using MetaExchangeSowaLabs.Core.Files;
-using MetaExchangeSowaLabs.Services.Services.Interfaces;
+using MetaExchangeSowaLabs.Core.Json;
+using MetaExchangeSowaLabs.Services.Interfaces;
 
 namespace MetaExchangeSowaLabs.Services.Services
 {
@@ -19,8 +20,8 @@ namespace MetaExchangeSowaLabs.Services.Services
                 throw new IllegalAmountOfBtcException(amountOfBtc);
 
                 //Deserialize the entities
-            var unorderedOrderBooks = IoFiles.DeserializeEntity<OrderBookEntity>(orderBooks);
-            var userBalance = IoFiles.DeserializeEntity<OrderBookBalanceEntity>(orderBooksUserBalance);
+            var unorderedOrderBooks = Json.DeserializeEntity<OrderBookEntity>(orderBooks);
+            var userBalance = Json.DeserializeEntity<OrderBookBalanceEntity>(orderBooksUserBalance);
 
 
             //Some starting edge-cases
